requirements.txt
discord.py
pymongo
python-dotenv
aiohttp
